+++ tools/db_bench_tool.cc
@@ -2925,6 +2925,8 @@ class Benchmark {
     }
     delete prefix_extractor_;
     if (cache_.get() != nullptr) {
+      // Clear cache reference first
+      open_options_.write_buffer_manager.reset();
       // this will leak, but we're shutting down so nobody cares
       cache_->DisownData();
     }